server-backend/ai_service/requirements.txt
fastapi
uvicorn
requests
pydantic
pydantic-settings
google-generativeai
python-dotenv
httpx